--- apps/rules/src/main.ts
@@ -1,1 +1,0 @@
-console.log('Hello World!');
