+++ apps/rules/src/tests/firestore/contracts.spec.ts
@@ -5,12 +5,6 @@
 } from '@firebase/rules-unit-testing';
 import { testFixture } from './fixtures/data';
 import { Firestore, initFirestoreApp } from '@blockframes/testing/firebase/functions';
-import { META_COLLECTION_NAME, MAINTENANCE_DOCUMENT_NAME } from '@blockframes/utils/maintenance';
-
-//Meta collection, for maintenance control.
-const metaDoc = `${META_COLLECTION_NAME}/${MAINTENANCE_DOCUMENT_NAME}`;
-testFixture[metaDoc] = {};
-testFixture[metaDoc].endedAt = true;
 
 //TODO: 4200
 describe.skip('Contracts Rules Tests', () => {
